convert start position to uint64

+++ PControl_Matlab/controller/PanelsController.m
@@ -1022,7 +1022,7 @@ classdef PanelsController < handle
                 pat.start = [];                
                 self.pullResponse();
                 for rsp_i = rsp
-                    pat.start = [pat.start strfind(self.iBuf, [rsp_i cmd])-1];
+                    pat.start = [pat.start uint64(strfind(self.iBuf, [rsp_i cmd])-1)];
                 end
                 if ~isempty(pat.start)
                     pat.end = pat.start + uint64(self.iBuf(pat.start));
